Fixed uploads directory replaced with CDN Path in settings.json

+++ src/System/IO/AttachLib.php
@@ -13,16 +13,20 @@ class AttachLib
 	}
 	public function delete($attach_id)
 	{
-		$up_id = (int)$attach_id;
-		$r = $this->app->db->query("DELETE FROM uploads WHERE up_id=$up_id;");
+		if (empty($this->app->settings->site['cdnpath']) || !is_dir($this->app->settings->site['cdnpath'])) {
+			return false;
+		}
+
+		$up_id = (int) $attach_id;
+		$r     = $this->app->db->query("DELETE FROM uploads WHERE up_id=$up_id;");
 		if ($r) {
 			try {
-				if (file_exists($this->app->root . "uploads/" . $up_id))
-					unlink($this->app->root . "uploads/" . $up_id);
-				if (file_exists($this->app->root . "uploads/" . $up_id . "_v"))
-					unlink($this->app->root . "uploads/" . $up_id . "_v");
-				if (file_exists($this->app->root . "uploads/" . $up_id . "_t"))
-					unlink($this->app->root . "uploads/" . $up_id . "_t");
+				if (file_exists($this->app->settings->site['cdnpath'] . DIRECTORY_SEPARATOR . $up_id))
+					unlink($this->app->settings->site['cdnpath'] . DIRECTORY_SEPARATOR . $up_id);
+				if (file_exists($this->app->settings->site['cdnpath'] . DIRECTORY_SEPARATOR . $up_id . "_v"))
+					unlink($this->app->settings->site['cdnpath'] . DIRECTORY_SEPARATOR . $up_id . "_v");
+				if (file_exists($this->app->settings->site['cdnpath'] . DIRECTORY_SEPARATOR . $up_id . "_t"))
+					unlink($this->app->settings->site['cdnpath'] . DIRECTORY_SEPARATOR . $up_id . "_t");
 			} catch (\Exception $e) {
 			}
 			return true;
@@ -33,8 +37,8 @@ class AttachLib
 
 	public function detach($attach_id)
 	{
-		$up_id = (int)$attach_id;
-		$r = $this->app->db->query("UPDATE uploads SET up_rel = NULL ,up_active = 0 WHERE up_id=$up_id;");
+		$up_id = (int) $attach_id;
+		$r     = $this->app->db->query("UPDATE uploads SET up_rel = NULL ,up_active = 0 WHERE up_id=$up_id;");
 		if ($r) {
 			return true;
 		} else {
